// gantryweb/Frontend/Gantry/src/app/common/models/racing-tags.model.ts
export enum FavouriteTags {
    favourite = "F",
    jointFavourite = "JF",
    jointSecondFavourite = "J2F",
    jointThirdFavourite = "J3F",
    combinedFavorite = "CF"
}

export enum ToteTags {
    forecast = "FC",
    tricast = "TC"
}

export enum RunnerType {
    Horse = "HORSE",
    Dog = "DOG",
    Motor = "MOTOR"
}